requirements.txt
python-dateutil~=2.8.2
requests~=2.27.1
mdutils~=1.3.1
pytz~=2021.3
feedparser~=6.0.8